--- crud/src/main/java/com/redhat/lightblue/hooks/HookPostParseListener.java
@@ -1,7 +1,0 @@
-package com.redhat.lightblue.hooks;
-
-public interface HookPostParseListener {
-
-    void fire(CRUDHook hook);
-
-}
